requirements.txt
altair==5.5.0
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
filetype==1.2.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.2.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.14.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.2.0
paypalrestsdk==1.13.3
pillow==10.2.0
plotly==5.24.1
protobuf==4.25.5
psycopg2-binary==2.9.9
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.49
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
streamlit==1.31.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
zipp==3.21.0
